pyproject.toml
[project]
name = "mas_lib"
version = "0.1.0"
description = "MAS Library"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "e2b-code-interpreter>=2.3.0",
    "fastmcp>=2.13.0.2",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-experimental>=0.0.42",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.2",
    "langchain-tavily>=0.2.13",
    "langgraph>=1.0.2",
    "marimo>=0.17.7",
    "markitdown[all]>=0.1.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["mas_lib*"]
